# Conflicts:
#	Common/ac/gamesetupstruct.cpp
#	Common/ac/gamesetupstruct.h
#	Common/game/main_game_file.cpp
#	Common/game/room_file.cpp
#	Engine/ac/game.cpp
#	Engine/ac/global_object.cpp
#	Engine/ac/room.cpp
#	Engine/ac/roomobject.h
#	Engine/game/game_init.cpp
#	Engine/game/savegame.cpp
#	Engine/game/savegame.h
#	Engine/game/savegame_components.cpp
#	Engine/game/savegame_components.h

+++ Engine/ac/dynamicsprite.cpp
@@ -447,7 +447,7 @@ void add_dynamic_sprite(int gotSlot, Bitmap *redin, bool hasAlpha) {
 
 void free_dynamic_sprite (int gotSlot) {
 
-  if ((gotSlot < 0) || (gotSlot >= spriteset.GetSpriteSlotCount()))
+  if ((gotSlot < 0) || ((size_t)gotSlot >= spriteset.GetSpriteSlotCount()))
     quit("!FreeDynamicSprite: invalid slot number");
 
   if ((game.SpriteInfos[gotSlot].Flags & SPF_DYNAMICALLOC) == 0)
